finished testing router
